+++ modules/default/calendar/calendarfetcher.js
@@ -3,6 +3,7 @@ const ical = require("node-ical");
 const Log = require("logger");
 const NodeHelper = require("node_helper");
 const CalendarFetcherUtils = require("./calendarfetcherutils");
+const { getUserAgent } = require("#server_functions");
 const { scheduleTimer } = require("#module_functions");
 
 /**
@@ -30,10 +31,9 @@ const CalendarFetcher = function (url, reloadInterval, excludedEvents, maximumEn
 	const fetchCalendar = () => {
 		clearTimeout(reloadTimer);
 		reloadTimer = null;
-		const nodeVersion = Number(process.version.match(/^v(\d+\.\d+)/)[1]);
 		let httpsAgent = null;
 		let headers = {
-			"User-Agent": `Mozilla/5.0 (Node.js ${nodeVersion}) MagicMirror/${global.version}`
+			"User-Agent": getUserAgent()
 		};
 
 		if (selfSignedCert) {